Miniexample Survival:
removed last consumer was left by mistake
improved code even further
set project to webgl
improving comments

+++ Example2-Unity Hybrid-Survival/Assets/Code/ECS/PlayerLayer/Engines/PlayerDeathEngine.cs
@@ -32,7 +32,7 @@ namespace Svelto.ECS.Example.Survive.Player
 
                     for (int i = 0; i < filteredIndices.count; i++)
                     {
-                        anim[filteredIndices[i]].animationState = new AnimationState((int)PlayerAnimations.Die);
+                        anim[filteredIndices[i]].animationState = new AnimationState(PlayerAnimations.Die);
                         sounds[filteredIndices[i]].playOneShot = (int)AudioType.death;
                         
                         var egid = new EGID(ids[filteredIndices[i]], group);
